Add ARS configuration options; Fix summarization bug
The bug was found when testing with the new ARS configuration. The bug occurs
when an ARA modifies the QG by adding additional nodes and edges. When one of
these new edges is chosen such that its object and subject are both variable
nodes the summarization code failed because it expected at least one node on
every edge to be static.

For now the workaround is to simply skip the entire answer if this is found.
An improvement in the future would be to skip this edge instead of the entire
answer.