Add curie-search endpoint config; Fix bug in evidence summarization
Found a bug in evidence summarization after running end to end testing with no
mock services. The bug occurs with an ARA modifies its QG to either go from
a 2 hop query to a >2 hop query or when it adds edges that change the predicate
direction. It either case it was possible for the evidence summarization code to
break. The temporary workaround is to omit all answers that contain QGs with >2
hops and to try to find any edge in the QG that matches with the primary
predicates instead of assuming all edges in the ARAs QG will match the original
QG.