Remove memoization as per gregr's advice
"If this is being run as part of a long-lived server, I recommend against using Racket's memo library, since it will hold on to whatever it computes forever.

If you really need memoization, you're better off managing it manually so that you can scope the life of the memo table more narrowly."

Given that this procedure is run once at startup, we can drop the memoization.