Only clone FE if not already present to avoid sh script errors
If 'nobuild' is provided as an arg to the script, it will not run the
npm build steps. This modification is currently made in a way that
doesn't require any changes to the existing usages of this script in
other places; in the future a little extra work should be put in to
treat these as proper args.