Update ARSClient to apply filters universally, not just when fetching full results.
This is currently required to match what the UI requires. There may be
a case for applying filters more judiciously than this, as we are now
losing some generality. Fetching data is the only expensive operation,
therefore it's arguable that our original behavior was more general
and that non-data filtering out of results should happen in another
layer, but this will complicate the code without immediate benefit. So
we're going to make filters more universally, and not worry about the
loss of generality for now.