Update store classes to take either a pool or connection params. Update Auth Service to make use of it.
This is important because certain services, the Auth Service being a
prime example, will eventually need to use transactions, which can't
be done without all queries sharing the client. Our current style of
classes still doesn't fully allow for this, as each query
independently requests a client, but this is a first step towards
making transactions possible.