Add controller classes for the query and config API, and for login/logout.
When an object method is passed as a callback, the parent object is no
longer available via `this`. The use of an explicit bind() call in the
router is how one works around this.

It's a little ugly, but it allows us to have controllers share access to
the services/stores across their handlers without resorting to worse
access patters like a singleton service locator, etc.