fix bug with merging node tags into paths
This issue occured because some facets only apply to result nodes on
a path and some facets apply to all nodes on a path. In the case
where an internal node of a path was also a result node the path
would incorrectly pick up result level tags from an internal node.